Create a Freesurfer qdec

Create a Freesurfer qdec from a model formula, utilising R's robust model formula syntax. The function also scales continuous variables, and can keep the original data in the output.

Usage

make_fs_qdec(data, formula, path = NULL, keep_orig = FALSE)

Arguments

data

data.frame, list or environment containing the data. Neither matrix nor an array will be accepted.

formula

a model formula or terms object. For simple qdec file, with binary columns for all levels of a factor, make sure to include a formula with -1 to remove the intercept.

path

an option file path to write qdec to csv

keep_orig

logical or vector of column names, to keep the original data in the output. Default is FALSE.

Value

data.frame with model matrix and scaled continuous variables.

Details

The formula should in all likelihood also include the -1 to remove the intercept, as this will provide a matrix where all levels of factor variables have their own binary column. This is necessary to follow the steps from the Freesurfer documentation.

It is highly recommended to have the base-id's for Freesurfer in their own column, and request to have this remain in the qdec by using the id-column's name in the keep_orig argument.

Examples

cars <- mtcars
cars$cyl <- as.factor(cars$cyl)
cars$gear <- as.factor(cars$gear)

make_fs_qdec(cars, mpg ~ cyl + hp)
#> # A data frame: 32 × 12
#>    `(Intercept)`  cyl6  cyl8   mpgz   dispz    hpz  dratz      wtz  qsecz    vsz
#>  *         <dbl> <dbl> <dbl>  <dbl>   <dbl>  <dbl>  <dbl>    <dbl>  <dbl>  <dbl>
#>  1             1     1     0  0.151 -0.571  -0.535  0.568 -0.610   -0.777 -0.868
#>  2             1     1     0  0.151 -0.571  -0.535  0.568 -0.350   -0.464 -0.868
#>  3             1     0     0  0.450 -0.990  -0.783  0.474 -0.917    0.426  1.12 
#>  4             1     1     0  0.217  0.220  -0.535 -0.966 -0.00230  0.890  1.12 
#>  5             1     0     1 -0.231  1.04    0.413 -0.835  0.228   -0.464 -0.868
#>  6             1     1     0 -0.330 -0.0462 -0.608 -1.56   0.248    1.33   1.12 
#>  7             1     0     1 -0.961  1.04    1.43  -0.723  0.361   -1.12  -0.868
#>  8             1     0     0  0.715 -0.678  -1.24   0.175 -0.0278   1.20   1.12 
#>  9             1     0     0  0.450 -0.726  -0.754  0.605 -0.0687   2.83   1.12 
#> 10             1     1     0 -0.148 -0.509  -0.345  0.605  0.228    0.253  1.12 
#> # ℹ 22 more rows
#> # ℹ 2 more variables: amz <dbl>, carbz <dbl>

# Remove the intercept, necessary to follow
# steps from Freesurfer docs
make_fs_qdec(cars, mpg ~ -1 + cyl + hp)
#> # A data frame: 32 × 12
#>     cyl4  cyl6  cyl8   mpgz   dispz    hpz  dratz      wtz  qsecz    vsz    amz
#>  * <dbl> <dbl> <dbl>  <dbl>   <dbl>  <dbl>  <dbl>    <dbl>  <dbl>  <dbl>  <dbl>
#>  1     0     1     0  0.151 -0.571  -0.535  0.568 -0.610   -0.777 -0.868  1.19 
#>  2     0     1     0  0.151 -0.571  -0.535  0.568 -0.350   -0.464 -0.868  1.19 
#>  3     1     0     0  0.450 -0.990  -0.783  0.474 -0.917    0.426  1.12   1.19 
#>  4     0     1     0  0.217  0.220  -0.535 -0.966 -0.00230  0.890  1.12  -0.814
#>  5     0     0     1 -0.231  1.04    0.413 -0.835  0.228   -0.464 -0.868 -0.814
#>  6     0     1     0 -0.330 -0.0462 -0.608 -1.56   0.248    1.33   1.12  -0.814
#>  7     0     0     1 -0.961  1.04    1.43  -0.723  0.361   -1.12  -0.868 -0.814
#>  8     1     0     0  0.715 -0.678  -1.24   0.175 -0.0278   1.20   1.12  -0.814
#>  9     1     0     0  0.450 -0.726  -0.754  0.605 -0.0687   2.83   1.12  -0.814
#> 10     0     1     0 -0.148 -0.509  -0.345  0.605  0.228    0.253  1.12  -0.814
#> # ℹ 22 more rows
#> # ℹ 1 more variable: carbz <dbl>
make_fs_qdec(cars, mpg ~ -1 + cyl + hp + gear)
#> # A data frame: 32 × 14
#>     cyl4  cyl6  cyl8 gear4 gear5   mpgz   dispz    hpz  dratz      wtz  qsecz
#>  * <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>   <dbl>  <dbl>  <dbl>    <dbl>  <dbl>
#>  1     0     1     0     1     0  0.151 -0.571  -0.535  0.568 -0.610   -0.777
#>  2     0     1     0     1     0  0.151 -0.571  -0.535  0.568 -0.350   -0.464
#>  3     1     0     0     1     0  0.450 -0.990  -0.783  0.474 -0.917    0.426
#>  4     0     1     0     0     0  0.217  0.220  -0.535 -0.966 -0.00230  0.890
#>  5     0     0     1     0     0 -0.231  1.04    0.413 -0.835  0.228   -0.464
#>  6     0     1     0     0     0 -0.330 -0.0462 -0.608 -1.56   0.248    1.33 
#>  7     0     0     1     0     0 -0.961  1.04    1.43  -0.723  0.361   -1.12 
#>  8     1     0     0     1     0  0.715 -0.678  -1.24   0.175 -0.0278   1.20 
#>  9     1     0     0     1     0  0.450 -0.726  -0.754  0.605 -0.0687   2.83 
#> 10     0     1     0     1     0 -0.148 -0.509  -0.345  0.605  0.228    0.253
#> # ℹ 22 more rows
#> # ℹ 3 more variables: vsz <dbl>, amz <dbl>, carbz <dbl>

# Keep the original data also in the output
make_fs_qdec(cars, mpg ~ -1 + cyl + hp, keep_orig = TRUE)
#>                     cyl4 cyl6 cyl8        mpgz       dispz         hpz
#> Mazda RX4              0    1    0  0.15088482 -0.57061982 -0.53509284
#> Mazda RX4 Wag          0    1    0  0.15088482 -0.57061982 -0.53509284
#> Datsun 710             1    0    0  0.44954345 -0.99018209 -0.78304046
#> Hornet 4 Drive         0    1    0  0.21725341  0.22009369 -0.53509284
#> Hornet Sportabout      0    0    1 -0.23073453  1.04308123  0.41294217
#> Valiant                0    1    0 -0.33028740 -0.04616698 -0.60801861
#> Duster 360             0    0    1 -0.96078893  1.04308123  1.43390296
#> Merc 240D              1    0    0  0.71501778 -0.67793094 -1.23518023
#> Merc 230               1    0    0  0.44954345 -0.72553512 -0.75387015
#> Merc 280               0    1    0 -0.14777380 -0.50929918 -0.34548584
#> Merc 280C              0    1    0 -0.38006384 -0.50929918 -0.34548584
#> Merc 450SE             0    0    1 -0.61235388  0.36371309  0.48586794
#> Merc 450SL             0    0    1 -0.46302456  0.36371309  0.48586794
#> Merc 450SLC            0    0    1 -0.81145962  0.36371309  0.48586794
#> Cadillac Fleetwood     0    0    1 -1.60788262  1.94675381  0.85049680
#> Lincoln Continental    0    0    1 -1.60788262  1.84993175  0.99634834
#> Chrysler Imperial      0    0    1 -0.89442035  1.68856165  1.21512565
#> Fiat 128               1    0    0  2.04238943 -1.22658929 -1.17683962
#> Honda Civic            1    0    0  1.71054652 -1.25079481 -1.38103178
#> Toyota Corolla         1    0    0  2.29127162 -1.28790993 -1.19142477
#> Toyota Corona          1    0    0  0.23384555 -0.89255318 -0.72469984
#> Dodge Challenger       0    0    1 -0.76168319  0.70420401  0.04831332
#> AMC Javelin            0    0    1 -0.81145962  0.59124494  0.04831332
#> Camaro Z28             0    0    1 -1.12671039  0.96239618  1.43390296
#> Pontiac Firebird       0    0    1 -0.14777380  1.36582144  0.41294217
#> Fiat X1-9              1    0    0  1.19619000 -1.22416874 -1.17683962
#> Porsche 914-2          1    0    0  0.98049211 -0.89093948 -0.81221077
#> Lotus Europa           1    0    0  1.71054652 -1.09426581 -0.49133738
#> Ford Pantera L         0    0    1 -0.71190675  0.97046468  1.71102089
#> Ferrari Dino           0    1    0 -0.06481307 -0.69164740  0.41294217
#> Maserati Bora          0    0    1 -0.84464392  0.56703942  2.74656682
#> Volvo 142E             1    0    0  0.21725341 -0.88529152 -0.54967799
#>                           dratz          wtz       qsecz        vsz        amz
#> Mazda RX4            0.56751369 -0.610399567 -0.77716515 -0.8680278  1.1899014
#> Mazda RX4 Wag        0.56751369 -0.349785269 -0.46378082 -0.8680278  1.1899014
#> Datsun 710           0.47399959 -0.917004624  0.42600682  1.1160357  1.1899014
#> Hornet 4 Drive      -0.96611753 -0.002299538  0.89048716  1.1160357 -0.8141431
#> Hornet Sportabout   -0.83519779  0.227654255 -0.46378082 -0.8680278 -0.8141431
#> Valiant             -1.56460776  0.248094592  1.32698675  1.1160357 -0.8141431
#> Duster 360          -0.72298087  0.360516446 -1.12412636 -0.8680278 -0.8141431
#> Merc 240D            0.17475447 -0.027849959  1.20387148  1.1160357 -0.8141431
#> Merc 230             0.60491932 -0.068730634  2.82675459  1.1160357 -0.8141431
#> Merc 280             0.60491932  0.227654255  0.25252621  1.1160357 -0.8141431
#> Merc 280C            0.60491932  0.227654255  0.58829513  1.1160357 -0.8141431
#> Merc 450SE          -0.98482035  0.871524874 -0.25112717 -0.8680278 -0.8141431
#> Merc 450SL          -0.98482035  0.524039143 -0.13920420 -0.8680278 -0.8141431
#> Merc 450SLC         -0.98482035  0.575139986  0.08464175 -0.8680278 -0.8141431
#> Cadillac Fleetwood  -1.24665983  2.077504765  0.07344945 -0.8680278 -0.8141431
#> Lincoln Continental -1.11574009  2.255335698 -0.01608893 -0.8680278 -0.8141431
#> Chrysler Imperial   -0.68557523  2.174596366 -0.23993487 -0.8680278 -0.8141431
#> Fiat 128             0.90416444 -1.039646647  0.90727560  1.1160357  1.1899014
#> Honda Civic          2.49390411 -1.637526508  0.37564148  1.1160357  1.1899014
#> Toyota Corolla       1.16600392 -1.412682800  1.14790999  1.1160357  1.1899014
#> Toyota Corona        0.19345729 -0.768812180  1.20946763  1.1160357 -0.8141431
#> Dodge Challenger    -1.56460776  0.309415603 -0.54772305 -0.8680278 -0.8141431
#> AMC Javelin         -0.83519779  0.222544170 -0.30708866 -0.8680278 -0.8141431
#> Camaro Z28           0.24956575  0.636460997 -1.36476075 -0.8680278 -0.8141431
#> Pontiac Firebird    -0.96611753  0.641571082 -0.44699237 -0.8680278 -0.8141431
#> Fiat X1-9            0.90416444 -1.310481114  0.58829513  1.1160357  1.1899014
#> Porsche 914-2        1.55876313 -1.100967659 -0.64285758 -0.8680278  1.1899014
#> Lotus Europa         0.32437703 -1.741772228 -0.53093460  1.1160357  1.1899014
#> Ford Pantera L       1.16600392 -0.048290296 -1.87401028 -0.8680278  1.1899014
#> Ferrari Dino         0.04383473 -0.457097039 -1.31439542 -0.8680278  1.1899014
#> Maserati Bora       -0.10578782  0.360516446 -1.81804880 -0.8680278  1.1899014
#> Volvo 142E           0.96027290 -0.446876870  0.42041067  1.1160357  1.1899014
#>                          carbz  mpg cyl  hp
#> Mazda RX4            0.7352031 21.0   6 110
#> Mazda RX4 Wag        0.7352031 21.0   6 110
#> Datsun 710          -1.1221521 22.8   4  93
#> Hornet 4 Drive      -1.1221521 21.4   6 110
#> Hornet Sportabout   -0.5030337 18.7   8 175
#> Valiant             -1.1221521 18.1   6 105
#> Duster 360           0.7352031 14.3   8 245
#> Merc 240D           -0.5030337 24.4   4  62
#> Merc 230            -0.5030337 22.8   4  95
#> Merc 280             0.7352031 19.2   6 123
#> Merc 280C            0.7352031 17.8   6 123
#> Merc 450SE           0.1160847 16.4   8 180
#> Merc 450SL           0.1160847 17.3   8 180
#> Merc 450SLC          0.1160847 15.2   8 180
#> Cadillac Fleetwood   0.7352031 10.4   8 205
#> Lincoln Continental  0.7352031 10.4   8 215
#> Chrysler Imperial    0.7352031 14.7   8 230
#> Fiat 128            -1.1221521 32.4   4  66
#> Honda Civic         -0.5030337 30.4   4  52
#> Toyota Corolla      -1.1221521 33.9   4  65
#> Toyota Corona       -1.1221521 21.5   4  97
#> Dodge Challenger    -0.5030337 15.5   8 150
#> AMC Javelin         -0.5030337 15.2   8 150
#> Camaro Z28           0.7352031 13.3   8 245
#> Pontiac Firebird    -0.5030337 19.2   8 175
#> Fiat X1-9           -1.1221521 27.3   4  66
#> Porsche 914-2       -0.5030337 26.0   4  91
#> Lotus Europa        -0.5030337 30.4   4 113
#> Ford Pantera L       0.7352031 15.8   8 264
#> Ferrari Dino         1.9734398 19.7   6 175
#> Maserati Bora        3.2116766 15.0   8 335
#> Volvo 142E          -0.5030337 21.4   4 109

# Keep the original data of specific columns
# Use a character vector
make_fs_qdec(cars, mpg ~ -1 + cyl + hp, keep_orig = c("mpg", "gear"))
#>                     cyl4 cyl6 cyl8        mpgz       dispz         hpz
#> Mazda RX4              0    1    0  0.15088482 -0.57061982 -0.53509284
#> Mazda RX4 Wag          0    1    0  0.15088482 -0.57061982 -0.53509284
#> Datsun 710             1    0    0  0.44954345 -0.99018209 -0.78304046
#> Hornet 4 Drive         0    1    0  0.21725341  0.22009369 -0.53509284
#> Hornet Sportabout      0    0    1 -0.23073453  1.04308123  0.41294217
#> Valiant                0    1    0 -0.33028740 -0.04616698 -0.60801861
#> Duster 360             0    0    1 -0.96078893  1.04308123  1.43390296
#> Merc 240D              1    0    0  0.71501778 -0.67793094 -1.23518023
#> Merc 230               1    0    0  0.44954345 -0.72553512 -0.75387015
#> Merc 280               0    1    0 -0.14777380 -0.50929918 -0.34548584
#> Merc 280C              0    1    0 -0.38006384 -0.50929918 -0.34548584
#> Merc 450SE             0    0    1 -0.61235388  0.36371309  0.48586794
#> Merc 450SL             0    0    1 -0.46302456  0.36371309  0.48586794
#> Merc 450SLC            0    0    1 -0.81145962  0.36371309  0.48586794
#> Cadillac Fleetwood     0    0    1 -1.60788262  1.94675381  0.85049680
#> Lincoln Continental    0    0    1 -1.60788262  1.84993175  0.99634834
#> Chrysler Imperial      0    0    1 -0.89442035  1.68856165  1.21512565
#> Fiat 128               1    0    0  2.04238943 -1.22658929 -1.17683962
#> Honda Civic            1    0    0  1.71054652 -1.25079481 -1.38103178
#> Toyota Corolla         1    0    0  2.29127162 -1.28790993 -1.19142477
#> Toyota Corona          1    0    0  0.23384555 -0.89255318 -0.72469984
#> Dodge Challenger       0    0    1 -0.76168319  0.70420401  0.04831332
#> AMC Javelin            0    0    1 -0.81145962  0.59124494  0.04831332
#> Camaro Z28             0    0    1 -1.12671039  0.96239618  1.43390296
#> Pontiac Firebird       0    0    1 -0.14777380  1.36582144  0.41294217
#> Fiat X1-9              1    0    0  1.19619000 -1.22416874 -1.17683962
#> Porsche 914-2          1    0    0  0.98049211 -0.89093948 -0.81221077
#> Lotus Europa           1    0    0  1.71054652 -1.09426581 -0.49133738
#> Ford Pantera L         0    0    1 -0.71190675  0.97046468  1.71102089
#> Ferrari Dino           0    1    0 -0.06481307 -0.69164740  0.41294217
#> Maserati Bora          0    0    1 -0.84464392  0.56703942  2.74656682
#> Volvo 142E             1    0    0  0.21725341 -0.88529152 -0.54967799
#>                           dratz          wtz       qsecz        vsz        amz
#> Mazda RX4            0.56751369 -0.610399567 -0.77716515 -0.8680278  1.1899014
#> Mazda RX4 Wag        0.56751369 -0.349785269 -0.46378082 -0.8680278  1.1899014
#> Datsun 710           0.47399959 -0.917004624  0.42600682  1.1160357  1.1899014
#> Hornet 4 Drive      -0.96611753 -0.002299538  0.89048716  1.1160357 -0.8141431
#> Hornet Sportabout   -0.83519779  0.227654255 -0.46378082 -0.8680278 -0.8141431
#> Valiant             -1.56460776  0.248094592  1.32698675  1.1160357 -0.8141431
#> Duster 360          -0.72298087  0.360516446 -1.12412636 -0.8680278 -0.8141431
#> Merc 240D            0.17475447 -0.027849959  1.20387148  1.1160357 -0.8141431
#> Merc 230             0.60491932 -0.068730634  2.82675459  1.1160357 -0.8141431
#> Merc 280             0.60491932  0.227654255  0.25252621  1.1160357 -0.8141431
#> Merc 280C            0.60491932  0.227654255  0.58829513  1.1160357 -0.8141431
#> Merc 450SE          -0.98482035  0.871524874 -0.25112717 -0.8680278 -0.8141431
#> Merc 450SL          -0.98482035  0.524039143 -0.13920420 -0.8680278 -0.8141431
#> Merc 450SLC         -0.98482035  0.575139986  0.08464175 -0.8680278 -0.8141431
#> Cadillac Fleetwood  -1.24665983  2.077504765  0.07344945 -0.8680278 -0.8141431
#> Lincoln Continental -1.11574009  2.255335698 -0.01608893 -0.8680278 -0.8141431
#> Chrysler Imperial   -0.68557523  2.174596366 -0.23993487 -0.8680278 -0.8141431
#> Fiat 128             0.90416444 -1.039646647  0.90727560  1.1160357  1.1899014
#> Honda Civic          2.49390411 -1.637526508  0.37564148  1.1160357  1.1899014
#> Toyota Corolla       1.16600392 -1.412682800  1.14790999  1.1160357  1.1899014
#> Toyota Corona        0.19345729 -0.768812180  1.20946763  1.1160357 -0.8141431
#> Dodge Challenger    -1.56460776  0.309415603 -0.54772305 -0.8680278 -0.8141431
#> AMC Javelin         -0.83519779  0.222544170 -0.30708866 -0.8680278 -0.8141431
#> Camaro Z28           0.24956575  0.636460997 -1.36476075 -0.8680278 -0.8141431
#> Pontiac Firebird    -0.96611753  0.641571082 -0.44699237 -0.8680278 -0.8141431
#> Fiat X1-9            0.90416444 -1.310481114  0.58829513  1.1160357  1.1899014
#> Porsche 914-2        1.55876313 -1.100967659 -0.64285758 -0.8680278  1.1899014
#> Lotus Europa         0.32437703 -1.741772228 -0.53093460  1.1160357  1.1899014
#> Ford Pantera L       1.16600392 -0.048290296 -1.87401028 -0.8680278  1.1899014
#> Ferrari Dino         0.04383473 -0.457097039 -1.31439542 -0.8680278  1.1899014
#> Maserati Bora       -0.10578782  0.360516446 -1.81804880 -0.8680278  1.1899014
#> Volvo 142E           0.96027290 -0.446876870  0.42041067  1.1160357  1.1899014
#>                          carbz  mpg gear
#> Mazda RX4            0.7352031 21.0    4
#> Mazda RX4 Wag        0.7352031 21.0    4
#> Datsun 710          -1.1221521 22.8    4
#> Hornet 4 Drive      -1.1221521 21.4    3
#> Hornet Sportabout   -0.5030337 18.7    3
#> Valiant             -1.1221521 18.1    3
#> Duster 360           0.7352031 14.3    3
#> Merc 240D           -0.5030337 24.4    4
#> Merc 230            -0.5030337 22.8    4
#> Merc 280             0.7352031 19.2    4
#> Merc 280C            0.7352031 17.8    4
#> Merc 450SE           0.1160847 16.4    3
#> Merc 450SL           0.1160847 17.3    3
#> Merc 450SLC          0.1160847 15.2    3
#> Cadillac Fleetwood   0.7352031 10.4    3
#> Lincoln Continental  0.7352031 10.4    3
#> Chrysler Imperial    0.7352031 14.7    3
#> Fiat 128            -1.1221521 32.4    4
#> Honda Civic         -0.5030337 30.4    4
#> Toyota Corolla      -1.1221521 33.9    4
#> Toyota Corona       -1.1221521 21.5    3
#> Dodge Challenger    -0.5030337 15.5    3
#> AMC Javelin         -0.5030337 15.2    3
#> Camaro Z28           0.7352031 13.3    3
#> Pontiac Firebird    -0.5030337 19.2    3
#> Fiat X1-9           -1.1221521 27.3    4
#> Porsche 914-2       -0.5030337 26.0    5
#> Lotus Europa        -0.5030337 30.4    5
#> Ford Pantera L       0.7352031 15.8    5
#> Ferrari Dino         1.9734398 19.7    5
#> Maserati Bora        3.2116766 15.0    5
#> Volvo 142E          -0.5030337 21.4    4